set OPENSCADPATH to fix relative includes in preview
Uses the environment variable approach suggested in
https://github.com/openscad/emacs-scad-mode/pull/3 to fix relative file
includes.

This sets the `OPENSCADPATH` environment variable to be `default-directory`.
There might be some faults in assuming we can just set this (perhaps other users
also set this), but at least this moves us forward in being able to preview and
make files using `include` with relative directories.

This has been tested successfully with `scad-preview`, but I'm unfamiliar with
`flymake` enough that I'm not sure how to properly test it.